// src/features/todolists/api/tasksApi.types.ts
import {TaskStatus,TaskPriority} from '@/common/enums'
import { z } from 'zod'

export const DomainTaskSchema = z.object({
    description: z.string().nullable(),
    title: z.string(),
    status: z.nativeEnum(TaskStatus),
    priority: z.nativeEnum(TaskPriority),
    startDate: z.string().nullable(),
    deadline: z.string().nullable(),
    id: z.string(),
    todoListId: z.string(),
    order: z.number(),
    addedDate: z.string().datetime({local: true}),
})

export type DomainTask = z.infer<typeof DomainTaskSchema>

export type GetTasksResponse = {
    error: string | null
    totalCount: number
    items: DomainTask[]
}

export type UpdateTaskModel = {
    description: string | null
    title: string
    status: number
    priority: number
    startDate: string | null
    deadline: string | null
}